Install Git
This change will help me get closer to having a self-hosting setup. I
want to be able to install RHEL using this config and then have all of
the tools that I need to start updating this config.

I could have just put git into the %packages section, but I decided to
instead use Ansible to install the package. If I want any of the changes
make to the kickstart file to apply to an already existing system, then
I’ll have to re-install that system. With an Ansible playbook, I’ll be
able to apply some changes without having to reinstall the system.

I could have used ansible-pull to retrieve the playbook. Doing so would
have simplified by kickstart file considerably. Doing so would also mean
that an Internet connection would be required to deploy the config.
(Although, now that I think about it, you might actually need an
Internet connection anyway. I’m not sure whether or not DNF commands run
during %post scripts use the installation DVD as a software source or
not.)