Allow pip --user packages to be safely upgraded
This change will make it easy for me to safely keep pip packages
up-to-date. The motivation behind this change is to help make
jasons-rhel-config self-hosting. I need to install the mypy pip package
to be able to do everything that I want to do with jasons-rhel-config on
a machine that was deployed using the config.

The key word here is “safely”. There’s two problems that could
potentially occur when updating pip packages:

1. Files might be overwritten while a running application is using them.
2. All of the packages might not get fully updated or incompatible
package versions might get used.

To solve the first problem, this commit creates a systemd target that
causes jayman’s pip packages to be updated. To solve the second problem,
this commit uses <http://github.com/achillesrasquinha/pipupgrade>.

One concern I have with using multiple package managers on one system is
updates. In the past, I’m sure that I’ve run “dnf upgrade” without
remembering to run “flatpak update”. To this end, this commit makes sure
that the process for updating pip packages is guarenteed to trigger the
process for updating RPMs.