Replace NEWLINE constant with literal "\n"
Originally, the NEWLINE constant was added to work around the fact that
you can’t use escape sequences in f-strings. Eventually, I stopped using
the NEWLINE variable in an f-string, but kept it around because it was
used elsewhere.

Before this change, NEWLINE was only being used in one place. This
change removes NEWLINE since writting "\n" is shorter and just as clear.