Create systemd target for setup and updates
If running “dnf update” inside a desktop is a bad idea [1], then running
“dnf install” inside a desktop is probably also a bad idea (installing
one package might update other packages). To work around this problem,
this commit creates a systemd target that starts just enough of the
operating system to run a shell script. That shell script will be
responsible for configuring and updating the system.

I could have just done the installation and configuration in a %post
script, but then I would have to reinstall the OS whenever I changed my
configuration. This commit allows me to simply run
post_script_payload/entry-point and reboot in order to test out a new
version of this repo.

[1]: <https://lwn.net/Articles/702629/>