Explicitly specify set-up-and-update dependencies
At the moment, this is pretty redundant since DNF is guaranteed to be
installed. That being said, at one point YUM was guaranteed to be
installed. At this point, YUM is just an alias for DNF, so it’s
reasonable to believe that YUM might not be guaranteed to be installed
in the future. For all we know, one day DNF will get replaced with
rpm-ostree.