Install Python
This will allow me to use a machine deployed with this config either to
deploy this config on another machine or to redeploy this config on
itself. Eating your own dogfood is important. Doing so ensures quality
because you have to live with any mistakes that you make which
encourages you to fix those mistakes.