Make hardware clock use UTC
Before this change, the hardware clock would be in local time. This
would cause a problem with virtual machines. The VM host would set the
guest’s virtual hardware clock to the current time in UTC when ever the
guest shut down or restarted. Then, the guest would boot, notice that
the time was wrong and fix it. This would cause output in TTYs and in
the system journal about logs getting rotated due to time jumping
backwards.