Prompt for hostname
Now that I’m trying to deploy the config on both a VM and on a laptop,
it’s much more important to be able to differentiate between different
systems by looking at their name. Specifically, I want to be able to
differentiate between systems on the Red Hat systems management page
[1]. Doing so is important since I’m only entitled to a maximum of 16
systems. I regularly have to go in and remove systems from that page, so
having each of them have different names will be useful.

[1]: <https://access.redhat.com/management/systems>