Install Inkscape
I’m currently working on a new Godot project, and I want to use the
latest dev snapshot of Godot (4.0 RC 1, at the moment [1]). I would use
it on my desktop, but my desktop runs NixOS. On NixOS, you genrally
can’t just download and run a binary [2]. At the moment, I’m working on
a nixpkgs [3] pull request that updates godot_4 to rc1, but it only
finished compiling recently.

The long and the short of it is: if you want to use the latest version
of Godot on NixOS, you typically have to wait for it to compile. While
I’m waiting for it to compile, I want to be able to continue to work on
Godot projects, so I’ve decided to use Godot on Jason-Lemur-Pro.

I started working on a Godot project, and I wanted to create an SVG file
for it. That’s why I decided to install Inkscape.

[1]: <https://godotengine.org/article/release-candidate-godot-4-0-rc-1/>
[2]: <https://nixos.wiki/wiki/Packaging/Binaries>
[3]: <https://github.com/NixOS/nixpkgs>