Add support for CentOS
I’ve been using a Red Hat Developer Subscriptions for Individuals in
order to get access to RHEL. According to the terms and conditions [1],
I’m only allowed to use this subscription as an individual working
independently. I recently got a job teaching programming. As a result of
those terms and conditions, I haven’t been using Jason-Lemur-Pro for
that job. Instead, I’ve been using my old laptop (which is falling
apart).

Now that this config supports CentOS [2], I’ll be able to install it on
my new laptop, and then I’ll be able to use that laptop for work.

[1]: <https://developers.redhat.com/terms-and-conditions>
[2]: <https://centos.org>