Simplify strategy for filtering device IDs
Before this change, set-up-and-update.sh would explictly check to see if
you were using Jason-Lemur-Pro or Jason-Lemur-Pro-VM-Test to prevent
them from adding their own Syncthing device IDs to themselves. The
problem with that behavior was that it allowed only two systems to run
this config: Jason-Lemur-Pro and Jason-Lemur-Pro-VM-Test. If a third
system was to be created, then I would have to add another clause to
that if statement to explicitly check if we’re running on that third
system.

This change makes the strategy for preventing a machine from adding its
own Syncthing device id simpler so that I don’t have to keep making that
if statement longer and longer. I’m making this change in preparation
for adding another test VM [1].

[1]: <mid:ddsk22j337xkgynbcubhmwetmjy4ynbf72yy2bhkrpfgxayqx6@bse6xs4wb6bl>